Update expressroute-locations.md
Last week, I proposed the same change. The business I added is similar to Bezeq International Ltd.(https://www.bezeqint.net/english) and Eurofiber (https://eurofiber.nl/microsoft-azure/). Last time, it was closed without a clear reason. However, I feel that this business is similar to the two mentioned above and that it should be included in the overview.